...but I do need to cast
git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@27842 bd178f5f-7bea-4014-844f-495ca19b8805

+++ src/org/starexec/util/ArchiveUtil.java
@@ -387,7 +387,7 @@ public class ArchiveUtil {
 	 * @return Unix file permissions
 	 */
 	private static int getUnixMode(File f) throws IOException {
-		return Files.getAttribute(f.toPath(), "unix:mode");
+		return (Integer) Files.getAttribute(f.toPath(), "unix:mode");
 	}
 
 	/**
